sebato: Norme ast_debug + norme is_script_complete
